+++ src/motor_drive.cpp
@@ -58,53 +58,53 @@ void command_motors() {
 
 }
 
-void superior_right_shoulder_abductor_cb(const std_msgs::Float32 &superior_right_shoulder_pos_msg) {
+void superior_right_shoulder_abductor_cb(const std_msgs::Float64 &superior_right_shoulder_pos_msg) {
     superior_right_shoulder_abductor_pwm = map2(superior_right_shoulder_pos_msg.data, -1.104, 1.2736, 160, 440) + superior_right_shoulder_abductor_offset;
     pwm.setPWM(0, 0, superior_right_shoulder_abductor_pwm);
 }
-void superior_left_shoulder_abductor_cb(const std_msgs::Float32 &superior_left_shoulder_pos_msg) {
+void superior_left_shoulder_abductor_cb(const std_msgs::Float64 &superior_left_shoulder_pos_msg) {
     superior_left_shoulder_abductor_pwm = map2(superior_left_shoulder_pos_msg.data, -1.104, 1.2736, 420, 140) + superior_left_shoulder_abductor_offset;
     pwm.setPWM(1, 0, superior_left_shoulder_abductor_pwm);
 }
-void inferior_right_shoulder_abductor_cb(const std_msgs::Float32 &inferior_right_shoulder_pos_msg) {
+void inferior_right_shoulder_abductor_cb(const std_msgs::Float64 &inferior_right_shoulder_pos_msg) {
     inferior_right_shoulder_abductor_pwm = map2(inferior_right_shoulder_pos_msg.data, -1.104, 1.2736, 420, 140) + inferior_right_shoulder_abductor_offset;
     pwm.setPWM(2, 0, inferior_right_shoulder_abductor_pwm);
 }
-void inferior_left_shoulder_abductor_cb(const std_msgs::Float32 &inferior_left_shoulder_pos_msg) {
+void inferior_left_shoulder_abductor_cb(const std_msgs::Float64 &inferior_left_shoulder_pos_msg) {
     inferior_left_shoulder_abductor_pwm = map2(inferior_left_shoulder_pos_msg.data, -1.104, 1.2736, 160, 440) + inferior_left_shoulder_abductor_offset;
     pwm.setPWM(3, 0, inferior_left_shoulder_abductor_pwm);
 }
 
-void superior_right_arm_extensor_cb(const std_msgs::Float32 &superior_right_arm_pos_msg) {
+void superior_right_arm_extensor_cb(const std_msgs::Float64 &superior_right_arm_pos_msg) {
     superior_right_arm_extensor_pwm = map2(superior_right_arm_pos_msg.data, -1.7831, 1.6133, 500, 100) + superior_right_arm_extensor_offset;
     pwm.setPWM(4, 0, superior_right_arm_extensor_pwm);
 }
-void superior_left_arm_extensor_cb(const std_msgs::Float32 &superior_left_arm_pos_msg) {
+void superior_left_arm_extensor_cb(const std_msgs::Float64 &superior_left_arm_pos_msg) {
     superior_left_arm_extensor_pwm = map2(superior_left_arm_pos_msg.data, -1.7831, 1.6133, 80, 480) + superior_left_arm_extensor_offset;
     pwm.setPWM(5, 0, superior_left_arm_extensor_pwm);
 }
-void inferior_right_arm_extensor_cb(const std_msgs::Float32 &inferior_right_arm_pos_msg) {
+void inferior_right_arm_extensor_cb(const std_msgs::Float64 &inferior_right_arm_pos_msg) {
     inferior_right_arm_extensor_pwm = map2(inferior_right_arm_pos_msg.data, -1.4434, 1.6982, 460, 90) + inferior_right_arm_extensor_offset;
     pwm.setPWM(6, 0, inferior_right_arm_extensor_pwm);
 }
-void inferior_left_arm_extensor_cb(const std_msgs::Float32 &inferior_left_arm_pos_msg) {
+void inferior_left_arm_extensor_cb(const std_msgs::Float64 &inferior_left_arm_pos_msg) {
     inferior_left_arm_extensor_pwm = map2(inferior_left_arm_pos_msg.data, -1.4434, 1.6982, 120, 490) + inferior_left_arm_extensor_offset;
     pwm.setPWM(7, 0, inferior_left_arm_extensor_pwm);
 }
 
-void superior_right_forearm_extensor_cb(const std_msgs::Float32 &superior_right_forearm_pos_msg) {
+void superior_right_forearm_extensor_cb(const std_msgs::Float64 &superior_right_forearm_pos_msg) {
     superior_right_forearm_extensor_pwm = map2(superior_right_forearm_pos_msg.data, -0.933987, 2.7171, 180, 610) + superior_right_forearm_extensor_offset;
     pwm.setPWM(8, 0, superior_right_forearm_extensor_pwm);
 }
-void superior_left_forearm_extensor_cb(const std_msgs::Float32 &superior_left_forearm_pos_msg) {
+void superior_left_forearm_extensor_cb(const std_msgs::Float64 &superior_left_forearm_pos_msg) {
     superior_left_forearm_extensor_pwm = map2(superior_left_forearm_pos_msg.data, 0, 2.80196, 290, -40) + superior_left_forearm_extensor_offset;
     pwm.setPWM(9, 0, superior_left_forearm_extensor_pwm);
 }
-void inferior_right_forearm_extensor_cb(const std_msgs::Float32 &inferior_right_forearm_pos_msg) {
+void inferior_right_forearm_extensor_cb(const std_msgs::Float64 &inferior_right_forearm_pos_msg) {
     inferior_right_forearm_extensor_pwm = map2(inferior_right_forearm_pos_msg.data, -0.2547, 2.8869, 260, 630) + inferior_right_forearm_extensor_offset;
     pwm.setPWM(10, 0, inferior_right_forearm_extensor_pwm);
 }
-void inferior_left_forearm_extensor_cb(const std_msgs::Float32 &inferior_left_forearm_pos_msg) {
+void inferior_left_forearm_extensor_cb(const std_msgs::Float64 &inferior_left_forearm_pos_msg) {
     inferior_left_forearm_extensor_pwm = map2(inferior_left_forearm_pos_msg.data, -0.33963, 2.46233, 330, 0) + inferior_left_forearm_extensor_offset;
     pwm.setPWM(11, 0, inferior_left_forearm_extensor_pwm);
 }
